pyproject.toml
[project]
name = "stemcell-agent"
version = "0.1.0"
description = "A scaffold for agents that specialize themselves for a domain, then freeze."
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.17",
    "openai>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.28",
]
